+++ app/utils/selectors.ts
@@ -1,36 +1,36 @@
+import { HN_SELECTORS } from '~app/constants';
+
 export const getControlPanelParentElement = (): HTMLElement | null => {
-  return document.querySelector(
-    'body > center > table > tbody > tr:nth-child(1) > td > table > tbody > tr > td:nth-child(3)',
-  );
+  return document.querySelector(HN_SELECTORS.CONTROL_PANEL_PARENT);
 };
 
 export const getTableBody = (): HTMLElement | null => {
-  return document.querySelector('body > center > table > tbody > tr:nth-child(3) > td > table > tbody');
+  return document.querySelector(HN_SELECTORS.TABLE_BODY);
 };
 
 export const getTitleRows = (tableBody: HTMLElement): NodeListOf<HTMLElement> => {
-  return tableBody.querySelectorAll('tr:nth-child(3n+1)');
+  return tableBody.querySelectorAll(HN_SELECTORS.TITLE_ROWS);
 };
 
 export const getInfoRows = (tableBody: HTMLElement): NodeListOf<HTMLElement> => {
-  return tableBody.querySelectorAll('tr:nth-child(3n+2)');
+  return tableBody.querySelectorAll(HN_SELECTORS.INFO_ROWS);
 };
 
 export const getSpacerRows = (tableBody: HTMLElement): NodeListOf<HTMLElement> => {
-  return tableBody.querySelectorAll('tr:nth-child(3n+3)');
+  return tableBody.querySelectorAll(HN_SELECTORS.SPACER_ROWS);
 };
 
 export const getPointsElement = (infoRow: HTMLElement): HTMLElement | null => {
-  return infoRow.querySelector('td.subtext > span > span.score');
+  return infoRow.querySelector(HN_SELECTORS.POINTS);
 };
 
 export const getTimeElement = (infoRow: HTMLElement): HTMLElement | null => {
-  const regularPost = infoRow.querySelector<HTMLElement>('td.subtext > span > span.age');
-  const promoPost = infoRow.querySelector<HTMLElement>('td.subtext > span.age');
+  const regularPost = infoRow.querySelector<HTMLElement>(HN_SELECTORS.TIME_REGULAR);
+  const promoPost = infoRow.querySelector<HTMLElement>(HN_SELECTORS.TIME_PROMO);
 
   return regularPost || promoPost;
 };
 
 export const getCommentsElement = (infoRow: HTMLElement): HTMLElement | null => {
-  return infoRow.querySelector('td.subtext > span > a[href^="item?id="]');
+  return infoRow.querySelector(HN_SELECTORS.COMMENTS);
 };